refactor(dashboard): move v2 panel stats to a perses_ file
All v2 code lives in perses_-prefixed files until the v1 code goes away.
Pure move of the stats block out of dashboard.go, tests alongside it.
